requirements.txt
langchain
langchain-core
langchain-community
langchain-experimental
langchain-ollama
langchain-chroma==0.2.5
chromadb-client==1.0.20
streamlit==1.48.0
pypdf==6.0.0
python-dotenv==1.1.1
PyYAML==6.0.1